Rename endpoint_mixins.py to mixins.py and update imports
